Fix format specifier at src/pet/petmem.c:1800:49
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@36550 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/pet/petmem.c
@@ -1797,7 +1797,7 @@ static int mem_dump_io(void *context, uint16_t addr)
                     petres.superpet_cpu_switch == SUPERPET_CPU_6502 ? "6502" :
                     petres.superpet_cpu_switch == SUPERPET_CPU_6809 ? "6809" :
                     "PROG (unimpl)");
-            mon_out("RAM write protect: $%x\n", spet_ramwp);
+            mon_out("RAM write protect: $%x\n", (unsigned int)spet_ramwp);
             mon_out("diagnostic sense: $%x\n", spet_diag);
             return 0;
         } else if (addr == 0xeffc) {
